+++ robotframework-ls/tests/robotframework_ls_tests/completions/test_auto_import.py
@@ -2,12 +2,16 @@ import pytest
 from typing import Dict, List
 
 
-def apply_completion(doc, completion):
+def apply_completion(doc, completion, expect_additional_text_edits=True):
     text_edit = completion["textEdit"]
     additional_text_edits = completion["additionalTextEdits"]
-    assert additional_text_edits
+    if expect_additional_text_edits:
+        assert additional_text_edits
+    else:
+        assert not additional_text_edits
     doc.apply_text_edits([text_edit])
-    doc.apply_text_edits(additional_text_edits)
+    if additional_text_edits:
+        doc.apply_text_edits(additional_text_edits)
 
 
 def test_completion_with_auto_import_basic_stdlib(
@@ -46,6 +50,127 @@ Library    Collections
 *** Test Cases ***
 User can call library
     Copy Dictionary"""
+    )
+
+
+def test_completion_with_auto_import_disabled(
+    workspace, cases, libspec_manager, workspace_dir
+):
+    from robotframework_ls.impl.completion_context import CompletionContext
+    from robotframework_ls.impl import auto_import_completions
+    from robotframework_ls.robot_config import RobotConfig
+    from robotframework_ls.impl.robot_generated_lsp_constants import (
+        OPTION_ROBOT_COMPLETIONS_KEYWORDS_NOT_IMPORTED_ENABLE,
+    )
+
+    cases.copy_to("case1", workspace_dir)
+    config = RobotConfig()
+    config.update({OPTION_ROBOT_COMPLETIONS_KEYWORDS_NOT_IMPORTED_ENABLE: False})
+
+    workspace.set_root(workspace_dir, libspec_manager=libspec_manager)
+    doc = workspace.put_doc("case1.robot")
+
+    doc.source = """
+*** Settings ***
+Library    case1_library
+
+*** Test Cases ***
+User can call library
+    Copy Diction"""
+
+    completions = auto_import_completions.complete(
+        CompletionContext(doc, workspace=workspace.ws, config=config), {}
+    )
+
+    assert len(completions) == 0
+
+
+def test_completion_with_auto_import_dont_add_import(
+    workspace, cases, libspec_manager, workspace_dir
+):
+    from robotframework_ls.impl.completion_context import CompletionContext
+    from robotframework_ls.impl import auto_import_completions
+    from robotframework_ls.robot_config import RobotConfig
+    from robotframework_ls.impl.robot_generated_lsp_constants import (
+        OPTION_ROBOT_COMPLETIONS_KEYWORDS_NOT_IMPORTED_ADD_IMPORT,
+    )
+
+    cases.copy_to("case1", workspace_dir)
+    config = RobotConfig()
+    config.update({OPTION_ROBOT_COMPLETIONS_KEYWORDS_NOT_IMPORTED_ADD_IMPORT: False})
+
+    workspace.set_root(workspace_dir, libspec_manager=libspec_manager)
+    doc = workspace.put_doc("case1.robot")
+
+    doc.source = """
+*** Settings ***
+Library    case1_library
+
+*** Test Cases ***
+User can call library
+    Copy Diction"""
+
+    completions = auto_import_completions.complete(
+        CompletionContext(doc, workspace=workspace.ws, config=config), {}
+    )
+
+    assert len(completions) == 1
+    apply_completion(doc, completions[0], expect_additional_text_edits=False)
+
+    assert (
+        doc.source
+        == """
+*** Settings ***
+Library    case1_library
+
+*** Test Cases ***
+User can call library
+    Copy Dictionary"""
+    )
+
+
+def test_completion_with_auto_import_dont_prefix_library(
+    workspace, cases, libspec_manager, workspace_dir
+):
+    from robotframework_ls.impl.completion_context import CompletionContext
+    from robotframework_ls.impl import auto_import_completions
+    from robotframework_ls.robot_config import RobotConfig
+    from robotframework_ls.impl.robot_generated_lsp_constants import (
+        OPTION_ROBOT_COMPLETIONS_KEYWORDS_NOT_IMPORTED_PREFIX_MODULE,
+    )
+
+    cases.copy_to("case1", workspace_dir)
+    config = RobotConfig()
+    config.update({OPTION_ROBOT_COMPLETIONS_KEYWORDS_NOT_IMPORTED_PREFIX_MODULE: True})
+
+    workspace.set_root(workspace_dir, libspec_manager=libspec_manager)
+    doc = workspace.put_doc("case1.robot")
+
+    doc.source = """
+*** Settings ***
+Library    case1_library
+
+*** Test Cases ***
+User can call library
+    Copy Diction"""
+
+    completions = auto_import_completions.complete(
+        CompletionContext(doc, workspace=workspace.ws, config=config), {}
+    )
+
+    assert len(completions) == 1
+    apply_completion(doc, completions[0])
+
+    assert (
+        doc.source
+        == """
+*** Settings ***
+Library    case1_library
+Library    Collections
+
+*** Test Cases ***
+User can call library
+    Collections.Copy Dictionary"""
     )