+++ robotframework-ls/docs/changelog.md
@@ -1,3 +1,21 @@
+NEXT
+-----------------------------
+
+### New features
+
+- `Robot Output` view is filled with contents of the run.
+
+- New Settings to configure completions from keywords not currently imported in the current module:
+
+  - `robot.completions.keywordsNotImported.enable`: Defines whether to show completions for keywords not currently imported (adds the proper import when applied).
+  - `robot.completions.keywordsNotImported.addImport`: Defines whether to actually add import for completions showing keywords not currently imported.
+  - `robot.completions.keywordsNotImported.prefixModule`: Defines whether completions showing keywords not currently imported should prefix completions with the module name.
+
+### Bugfixes
+
+- Colors in the `Interactive Console` could be wrong on some cases.
+
+
 New in 1.5.0 (2022-11-09)
 -----------------------------